contacts_app/requirements.txt
asn1crypto==0.22.0
cffi==1.10.0
cryptography==1.8.1
Django==1.11.2
django-cors-headers==2.1.0
djangorestframework==3.6.3
idna==2.5
olefile==0.44
packaging==16.8
Pillow==4.1.1
pycosat==0.6.2
pycparser==2.17
pyOpenSSL==17.0.0
pyparsing==2.1.4
pytz==2017.2
requests==2.14.2
six==1.10.0
